Convert to letter game, from number game
Using a similar structure as the previous number version, I have used chars to allocate lower case letters into a vector. In the future I will look to create a user defined class so that only one vector is needed, as opposed to one for answers and one for guesses.